Bots - Messaging
Refactored bots to allow them to process tasks on a messaging queue (currently rabbit).
Each entity that the bot sees gets sent to a different exchange

+++ code/agents/twitter/models.py
@@ -1,10 +1,13 @@
 from __future__ import annotations
-
-from dataclasses import dataclass
-from typing import Dict, Any
 import abc
-from utils import snowflake_time
+from dataclasses import dataclass
+from typing import Dict, Any, Optional
+
 import tweepy
+
+from utils import snowflake_time
+
+__all__ = ["Colour", "Color", "BaseModel", "User", "Tweet"]
 
 
 class Colour:
@@ -56,8 +59,10 @@ class Colour:
     def from_rgb(cls, r, g, b) -> Colour:
         """Factory method for a :class:`Colour` from an RGB tuple."""
         return cls((r << 16) + (g << 8) + b)
+
     def to_json(self):
         return str(self)
+
 
 Color = Colour
 
@@ -86,6 +91,10 @@ class BaseModel(abc.ABC):
     @abc.abstractmethod
     def from_json(json):
         """Returns an instance of this Model from the JSON supplied"""
+        raise NotImplementedError
+
+    @abc.abstractmethod
+    def to_json(self):
         raise NotImplementedError
 
 
@@ -136,6 +145,7 @@ class User(BaseModel):
         Returns the JSON that represents this entity
         """
         return self._json
+
     @staticmethod
     def from_json(api, json: dict):
         """Returns an instance of this Model from the JSON supplied"""
@@ -144,9 +154,10 @@ class User(BaseModel):
         # removing unused attributes
         ignore_attributes = ["id_str", "entities", "created_at", "utc_offset", "time_zone",
                              "geo_enabled", "is_translator", "is_translation_enabled",
-                             "needs_phone_verification", "translator_type", "status", "profile_banner_url"]
+                             "needs_phone_verification", "translator_type", "status",
+                             "profile_banner_url"]
         for i in ignore_attributes:
-            unclean_json.pop(i,None)
+            unclean_json.pop(i, None)
         # To make things easier for to us, we're going transforming some of the attributes into objects
         # TODO: Currently only for colours, do the same for entities
         unclean_json["profile_background_color"] = Colour.from_hex(
@@ -172,11 +183,85 @@ class User(BaseModel):
         return self._api.followers(user_id=self.id, **kargs)
 
     def follow(self):
+        if self.following:
+            return
         self._api.create_friendship(user_id=self.id)
         self.following = True
         self._json["following"] = True
 
     def unfollow(self):
+        if not self.following:
+            return
         self._api.destroy_friendship(user_id=self.id)
         self.following = False
         self._json["following"] = False
+
+
+@dataclass
+class Tweet(BaseModel):
+    id: int
+    text: str
+
+    @property
+    def created_at(self):
+        """Returns the model's creation time in UTC."""
+        return str(snowflake_time(self.id))
+
+    truncated: str
+    entities: Dict
+    source: str
+    in_reply_to_status_id: Optional[int]
+    in_reply_to_user_id: Optional[int]
+    in_reply_to_screen_name: Optional[str]
+    user: User
+    is_quote_status: bool
+    retweet_count: int
+    favorite_count: int
+    favorited: bool
+    retweeted: bool
+    possibly_sensitive: bool
+    lang: str
+
+    def to_json(self):
+        unclean = self._json
+        # we don't need to send these 2 attribute
+        unclean.pop("favorited", None)
+        unclean.pop("retweeted", None)
+        # and also cleaning up the tweet object, by not filling in with the user
+        unclean["user"] = unclean["user"].id
+        return unclean
+
+    @staticmethod
+    def from_json(api, json: dict):
+        """Returns an instance of this Model from the JSON supplied"""
+        # copying the
+        unclean_json = json
+        # removing unused attributes
+        ignore_attributes = ["id_str", "created_at", "in_reply_to_status_id_str",
+                             "in_reply_to_user_id_str", "geo", "coordinates", "place",
+                             "contributors"]
+        for i in ignore_attributes:
+            unclean_json.pop(i, None)
+        # Filling in with User object for handiness
+        unclean_json["user"] = User.from_json(api, unclean_json["user"])
+        return Tweet(_api=api, _json=unclean_json, **unclean_json)
+
+    def retweet(self):
+        if self.retweeted:
+            return
+        self._api.retweet(self.id)
+        self.retweeted = True
+
+    def like(self):
+        # Unnecessary to make the API call in this case
+        if self.favorited:
+            return
+        self._api.create_favorite(self.id)
+        self.favorited = True
+
+    def dislike(self):
+        # Unnecessary to make the API call in this case
+        if not self.favorited:
+            return
+        self._api.destroy_favorite(self.id)
+        self.favorited = False
